Abortion(vcdExtra)	R Documentation

Abortion Opinion Data

Description

Opinions about abortion classified by gender and SES

Usage


data(Abortion)

Format

A 3-dimensional array resulting from cross-tabulating 3 variables for 1100 observations. The variable names and their levels are:

No 	 Name 	 Levels

1 	 Sex 	 "Female", "Male"

2 	 Status 	 "Lo", "Hi"

3 	 Support_Abortion 	 "Yes", "No"

Details

The combinations of Sex and Status represent four independent samples, having fixed Sex-Status marginal totals. Thus the Sex:Status association must be included in any loglinear model. Support_Abortion is a natural response variable.

Source

Christensen, R. (1990). Log-Linear Models, New York, NY: Springer-Verlag, p. 92, Example 3.5.2.

Christensen, R. (1997). Log-Linear Models and Logistic Regression, New York, NY: Springer, p. 100, Example 3.5.2.

Examples


data(Abortion)

# example goes here
ftable(Abortion)
mosaic(Abortion, shade=TRUE)

# stratified by Sex
fourfold(aperm(Abortion, 3:1))
# stratified by Status
fourfold(aperm(Abortion, c(3,1,2)))


[Package vcdExtra version 0.4-1 Index]