Bartlett(vcdExtra)	R Documentation

Bartlett data on plum root cuttings

Description

In an experiment to investigate the effect of cuitting length (two levels) and planting time (two levels) on the survival of plum root cuttings, 240 cuttings were planted for each of the 2 x 2 combinations of these factors, and their survival was later recorded. Bartlett (1935) used these data to illustrate a method for testing for no three-way interaction in a contingency table.

Usage


data(Bartlett)

Format

A 3-dimensional array resulting from cross-tabulating 3 variables for 960 observations. The variable names and their levels are:

No 	 Name 	 Levels

1 	 Alive 	 "Alive", "Dead"

2 	 Time 	 "Now", "Spring"

3 	 Length 	 "Long", "Short"

Source

Hand, D. and Daly, F. and Lunn, A. D.and McConway, K. J. and Ostrowski, E. (1994). A Handbook of Small Data Sets. London: Chapman & Hall, p. 15, # 19.

References

Bartlett, M. S. (1935). Contingency Table Interactions Journal of the Royal Statistical Society, Supplement, 1935, 2, 248-252.

Examples


data(Bartlett)

# example goes here
fourfold(Bartlett, mfrow=c(1,2))

mosaic(Bartlett, shade=TRUE)


[Package vcdExtra version 0.4-0 Index]