Dyke(vcdExtra)	R Documentation

Sources of knowledge of cancer

Description

Observational data on a sample of 1729 individuals, cross-classified in a 2^5 table according to their sources of information (read newspapers, listen to the radio, do 'solid' reading, attend lectures) and whether they have good or poor knowledge regarding cancer. Knowledge of cancer is often treated as the response.

Usage


data(Dyke)

Format

A 5-dimensional array resulting from cross-tabulating 5 variables for 1729 observations. The variable names and their levels are:

No 	 Name 	 Levels

1 	 Knowledge 	 "Good", "Poor"

2 	 Reading 	 "No", "Yes"

3 	 Radio 	 "No", "Yes"

4 	 Lectures 	 "No", "Yes"

5 	 Newspaper 	 "No", "Yes"

Source

Fienberg, S. E. (1980). The Analysis of Cross-Classified Categorical Data Cambridge, MA: MIT Press, p. 85, Table 5-6.

References

Dyke, G. V. and Patterson, H. D. (1952). Analysis of factorial arrangements when the data are proportions. Biometrics, 8, 1-12.

Lindsey, J. K. (1993). Models for Repeated Measurements Oxford, UK: Oxford University Press, p. 57.

Examples


data(Dyke)

# independence model
mosaic(Dyke, shade=TRUE)

# baseline model, Knowledge as response
dyke.mod0 <- loglm(~ Knowledge + (Reading * Radio * Lectures * Newspaper), data=Dyke)
dyke.mod0

mosaic(dyke.mod0)


[Package vcdExtra version 0.4-1 Index]