ICU(vcdExtra)	R Documentation

Death in the ICU

Description

The ICU data set consists of a sample of 200 subjects who were part of a much larger study on survival of patients following admission to an adult intensive care unit (ICU). The major goal of this study was to develop a logistic regression model to predict the probability of survival to hospital discharge of these patients and to study the risk factors associated with ICU mortality.

Usage

data(ICU)

Format

A data frame with 200 observations on the following 21 variables.

id

Patient ID

died

Died before discharge: a factor with levels No Yes

age

Age: a numeric vector

sex

Sex: a factor with levels Female Male

race

Race: a factor with levels Black Other White

service

Service at admission: a factor with levels Medical Surgery

cancer

Cancer part of problem?: a factor with levels No Yes

renal

History of chronic renal?: a factor with levels No Yes

infect

Infection probable?: a factor with levels No Yes

cpr

CPR prior to ICU admission?: a factor with levels No Yes

systolic

Systolic blood pressure: a numeric vector

hrtrate

Heart rate: a numeric vector

previcu

Previous admit to ICU?: a factor with levels No Yes

admit

Type of admission: a factor with levels Elective Emergency

fracture

Fracture?: a factor with levels No Yes

po2

PO2 inital blood gas: a numeric vector

ph

pH inital blood gas: a factor with levels <7.25 >=7.25

pco

PCO2 inital blood gas: a factor with levels <=45 >45

bic

Bicarbonate inital blood: a numeric vector

creatin

Creatinine inital blood: a factor with levels <=2 >2

coma

Consciousness at ICU: an ordered factor with levels None Stupor Coma

Details

Data were collected at Baystate Medical Center in Springfield, Massachusetts. The clinical aspects of this study are described in Lemeshow, Teres, Avrunin, and Pastides (1988).

Source

Hosmer and Lemeshow, Applied Logistic Regression, Wiley, (1989).

Lemeshow, S., Teres, D., Avrunin, J. S., Pastides, H. (1988). Predicting the Outcome of Intensive Care Unit Patients. Journal of the American Statistical Association, 83, 348-356.

References

Friendly, M. Visualizing Categorical Data, Cary, NC: SAS Institute, 2000, Appendix B.4.

Examples


data(ICU)
## maybe str(ICU) ; plot(ICU) ...

[Package vcdExtra version 0.2 Index]