vcdExtra-package(vcdExtra)	R Documentation

Extensions and additions to vcd: Visualizing Categorical Data

Description

This package provides additional data sets, documentation, and a few functions designed to extend the vcd package.

Details

Package: 	 vcdExtra

Type: 	 Package

Version: 	 0.4-2

Date: 	 2010-03-09

License: 	 GPL version 2 or newer

LazyLoad: 	 yes

The main purpose of this package is to serve as a sandbox for introducing extensions of mosaic plots that apply to loglinear models fitted using glm() and related, generalized nonlinear models fitted with gnm() in the gnm-package package. The method mosaic.glm is an initial implementation, intended to extend the mosaic.loglm method in the vcd package to a wider class of models.

In addition, there are several new data sets, a new vignette,

vcd-tutorial

Working with categorical data with R and the vcd package, vignette("vcd-tutorial", package = "vcdExtra")

and a few functions for manipulating categorical data sets.

Some of these extensions may be migrated into vcd.

Author(s)

Michael Friendly

Maintainer: Michael Friendly <friendly AT yorku.ca>

References

Friendly, M. Visualizing Categorical Data, Cary NC: SAS Insitute, 2000. Web materials: http://www.math.yorku.ca/SCS/vcd/.

Meyer, D.; Zeileis, A. & Hornik, K. The Strucplot Framework: Visualizing Multi-way Contingency Tables with vcd Journal of Statistical Software, 2006, 17, 1-48. Available in R via vignette("strucplot", package = "vcd")

Turner, H. and Firth, D. Generalized nonlinear models in R: An overview of the gnm package, 2007, http://eprints.ncrm.ac.uk/472/. Available in R via vignette("gnmOverview", package = "gnm").

See Also

gnm-package,mosaic

[Package vcdExtra version 0.4-2 Index]