GKgamma(vcdExtra)	R Documentation

Calculate Goodman-Kruskal Gamma for ordered tables

Description

The Goodman-Kruskal gamma statistic is a measure of association for ordinal factors in a two-way table proposed by Goodman and Kruskal (1954).

Usage


GKgamma(x, level = 0.95)


Arguments

x 	 A two-way frequency table, in matrix or table form. The rows and columns are considered to be ordinal factors

level 	 Confidence level for a significance test of gamma !=0

Value

Returns an object of class "GKgamma" with 6 components, as follows

gamma 	 The gamma statistic

C 	 Total number of concordant pairs in the table

D 	 Total number of disconcordant pairs in the table

sigma 	 Standard error of gamma

CIlevel 	 Confidence level

CI 	 Confidence interval

Author(s)

Michael Friendly; original version by Laura Thompson

References

Agresti, A. Categorical Data Analysis. John Wiley & Sons, 2002, pp. 57–59.

Goodman, L. A., & Kruskal, W. H. (1954). Measures of association for cross classifications. Journal of the American Statistical Association, 49, 732-764.

Goodman, L. A., & Kruskal, W. H. (1963). Measures of association for cross classifications III: Approximate sampling theory. Journal of the American Statistical Association, 58, 310-364.

See Also

assocstats, Kappa

Examples


data(JobSat)
GKgamma(JobSat)

[Package vcdExtra version 0.3-4 Index]