Vietnam(vcdExtra)	R Documentation

Student opinion about the war in Vietnam

Description

A survey of student opinion on the Vietnam War was taken at the University of North Carolina at Chapel Hill in May 1967 and published in the student newspaper. Students were asked to fill in ballot papers stating which policy out of A,B,C or D they supported. Responses were cross-classified by gender/year.

The result is a 2 x 5 x 4 contingency table, with factors sex, yr and resp.

Usage

data(Vietnam)

Format

A data frame in frequency form with 40 observations on the following 4 variables.

sex

a factor with levels F M

yr

an ordered factor with levels Fresh < Soph < Jr < Sr < Grad

resp

a factor with levels A B C D

Freq

a numeric vector

Details

The response categories can be considered ordered in terms of attitude about the war in Vietnam:

A

defeat North Vietnam by widespread bombing and land invasion

B

follow the present policy

C

withdraw troops to strong points and open negotiations on elections involving the Viet Cong

D

immediate withdrawal of all U.S. troops

Source

Aitkin, M. A., Anderson, D., Francis, B., & Hinde, J. Statistical Modelling in GLIM, Clarendon Press, 1989

References

Friendly, M. Visualizing Categorical Data, Cary, NC: SAS Institute, 2000, Appendix B.12.

Examples


data(Vietnam)
str(Vietnam) 
# make it into a table
Vietnam.table <- xtabs(Freq ~ sex+yr+resp, data=Vietnam)
structable(Vietnam.table)

[Package vcdExtra version 0.1 Index]